% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{freq}
\alias{freq}
\title{Computing frequencies of elements in a vector.}
\usage{
freq(x, elts)
}
\arguments{
\item{x}{A numeric vector that we wish to check how many times elements specified in \code{elts} appear.}

\item{elts}{A vector with elements that we wish to count in vector \code{x}.}
}
\value{
A numeric vector of the same length as \code{elts}, with names elements.
}
\description{
Computes number of occurrences of specified elements in a vector. This function behaves slightly different than R base function \code{\link{table}}, since we specify the elements, frequency of elements that were not found will be reported as zero. See also http://stackoverflow.com/questions/1923273/counting-the-number-of-elements-with-the-values-of-x-in-a-vector
}
\examples{
# create some test data
x <- rep(1:3, 4)
elts <- c(1, 2, 4)

# get the counts of specified elements 
freq(x, elts)
}
\seealso{
\code{\link{table}} which is very similar
}

