% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/other.R
\name{\%!in\%}
\alias{\%!in\%}
\title{A shortcut for negation of the \%in\% operator.}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{A data structure whose elements will be checked whether they are also present in data structure \code{y}.}

\item{y}{A data structure that will be checked whether elements from \code{x} are present in it.}
}
\value{
A logical vector of equal length as \code{x}, with element(s) not being present in \code{y} as TRUE. If \code{x} was a matrix, then the output corresponds to \code{x} converted to a vector.
}
\description{
Using a negation of \%in\% operator in the standard format produces less readable code. More natural way of using negation in this context would be to prepend 'in' with standard negation symbol, '!', which then results in using shorter and more readable 'x \%!in\% y' instead of '!(x \%in\% y)'.
}
\examples{
# create some numeric vectors
x <- 1:4
y <- 3:6

# standard form
results <- !(x \%in\% y)

# new abbreviated format
resultsNew <- x \%!in\% y
all(results == resultsNew)
}

