% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{idxMax}
\alias{idxMax}
\title{Get the position of a max element, resolving the ties in a reliable way.}
\usage{
idxMax(x, tol = 1e-16)
}
\arguments{
\item{x}{A numeric vector.}

\item{tol}{A scalar, small value, required to test the equality of real numbers.}
}
\value{
An integer denoting the position of a max element in vector \code{x}.
}
\description{
If all elements are the same or if there is more than one max element, max element is chosen randomly. At the moment it cannot deal with NA's, an error will be issued.
}
\examples{
# create some numeric vectors
x <- rep(1, 10)
y <- 1:10
z <- c(0, rep(1, 10))

# get the index of max element
set.seed(1234)
idxMax(x)  # repeating will give different results
idxMax(y)
idxMax(z)  # repeating will give different results
}

