% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{is.tri}
\alias{is.tri}
\title{Verify whether a matrix is a triangular matrix.}
\usage{
is.tri(L, method = lower.tri, tol = 1e-08)
}
\arguments{
\item{L}{A numeric matrix.}

\item{method}{Function that extracts either a lower or upper triangular part of the matrix.}

\item{tol}{Tolerance for determining the differences between the elements of the matrix, required for real numbers.}
}
\value{
A logical value, whether L is a triangular matrix or not.
}
\description{
Verify whether a matrix is a triangular matrix.
}
\examples{
# create a non-triangular matrix
mat <- matrix(runif(9), 3, 3)

# is it a triangular matrix?
is.tri(mat)
}

