% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mod}
\alias{mod}
\title{Computes a mod statistic.}
\usage{
mod(x, freq = FALSE, ties = "unbroken", na.rm = TRUE)
}
\arguments{
\item{x}{A numeric, logical or a character vector where the mod needs to be found.}

\item{freq}{A logical that indicates whether frequency of elements in the mod should be returned as well. By default set to FALSE.}

\item{ties}{A string that indicates how the ties should be broken. Possible values are: "unbroken" in which case ties are not broken and all of the tied elements are returned in the output (in this case, if \code{freq} is set to TRUE, frequency output will be a scalar that indicates frequency of all tied elements), and "first" where only the first element is returned. By default set to "unbroken".}

\item{na.rm}{A logical that indicates whether NA elements should be removed or not. By default set to TRUE.}
}
\value{
If \code{freq} set to FALSE, the output is a vector of the same class as \code{x} with a single element, indicating the mod of the vector provided in \code{x}. If \code{freq} set to FALSE, the output is a list with first element being the mod and second element being the frequency of the mod.
}
\description{
A function for finding the most frequent element in a vector and optionally its frequency. The intended use is not for estimating mod statistic, but for finding a mod in sequences where elements repeat themselves often. Hence, it should not be used with vectors of real numbers, as such elements are likely to be unique. Note that the function is not capable of breaking ties randomly.
}
\examples{
# create some numeric vectors
x <- c(1,1,1,2,2)
y <- c(1,1,2,2)

# no ties, with and without reporting frequency
mod(x)
mod(x, freq = TRUE)

# ties, with and without reporting frequency
mod(y)
mod(y, freq = TRUE)
}

