% Generated by roxygen2 (4.0.2): do not edit by hand
\name{repvec}
\alias{repvec}
\title{A function that replicates a vector and creates a matrix where each vector is either a row or a column.}
\usage{
repvec(vec, nrep, dim)
}
\arguments{
\item{vec}{A vector of any mode.}

\item{nrep}{A strictly positive scalar that determines how many times \code{vec} will be replicated and added as new rows or new columns, as indicated by \code{dim} argument.}

\item{dim}{A scalar with value of either 1 or 2. Value 1 will copy vector \code{vec} as rows, while value 2 will copy it as columns.}
}
\value{
A matrix with \code{length(vec)} rows and \code{nrep} columns if \code{dim==2} and vice versa if \code{dim==1}.
}
\description{
A function that replicates a vector and creates a matrix where each vector is either a row or a column.
}
\details{
A simple wrapper for \code{\link{rep}} that takes a vector and creates a matrix by replicating it certain number of times either as rows or as columns.
}
\examples{
vec <- c(1,2,3,4)
repvec(vec, 4, 2)
vec <- c("character", "vector")
repvec(vec, 3, 1)
vec <- c("and", "logical", "vector")
repvec(vec, 3, 2)
}
\seealso{
\code{\link{rep}}
}

