% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rescale}
\alias{rescale}
\title{Rescaling a numerical vector.}
\usage{
rescale(x, lower = 0, upper = 1, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector where numbers need to be rescaled.}

\item{lower}{Desired minimum value, by deafult set to 0.}

\item{upper}{Desired maximum value, by deafult set to 1.}

\item{na.rm}{A logical that indicates whether NA elements should be removed from the output or not. By default set to FALSE. Note that, if it is set to FALSE, the rescale will still remove NA's to do the computations, but it will simply leave the NA elements in the resulting vector.}
}
\value{
A numeric vector with rescaled elements from \code{x}.
}
\description{
A function for rescaling a vector of integers or real numbers to an arbitrary interval. Element in the original vector with the minimum value will be transformed to a number deteremined by  \code{lower} argument (zero by default), while the element with the maximum value will be transformed to  a number deteremined by \code{upper} argument (one by default). All other elements will be rescaled proportionally.
}
\examples{
# create some numeric vectors
x <- 1:10

# min element will be transformed to 0 and max element to 1
rescale(x)

# arbitrary lower and upper value
rescale(x, 20, 100)
}

