% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matlab.R
\name{rmat}
\alias{rmat}
\title{Simple matrix wrapper for gnerating random numbers.}
\usage{
rmat(distribution, nrow, ncol, byrow = TRUE, ...)
}
\arguments{
\item{distribution}{A distribution available in R that should generate the numbers for filling out the matrix, e.g. \code{\link{rnorm}} or code{\link{runif}}.}

\item{nrow}{A scalar indicating the desired number of rows.}

\item{ncol}{A scalar indicating the desired number of columns.}

\item{byrow}{A logical value, if TRUE (the default) the matrix is filled by rows, otherwise the matrix is filled by columns.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
A matrix with \code{nrow} rows and \code{ncol} columns, filled with numbers generated from \code{distribution}.
}
\description{
A shortcut for creating a matrix with randomly drawn numbers from arbitrary distribution available in R.
}
\examples{
# create a 2 by 2 matrix with numbers randomly drawn from uniform distribution
rmat(runif, 2, 2)

# same, but now we pass on additional arguments to runif
rmat(runif, 2, 2, min=10, max=20)
}
\seealso{
\code{\link{matrix}}, \code{\link{rnorm}}, \code{\link{runif}} etc.
}

