% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{sqrtmat}
\alias{sqrtmat}
\title{Computes square root of a matrix.}
\usage{
sqrtmat(mat, maxit = 50, tolerance = NA, verbose = FALSE)
}
\arguments{
\item{mat}{A numeric matrix where numbers need to be rescaled.}

\item{maxit}{Number of iterations of the algorithm, by deafult set to 50.}

\item{tolerance}{Tolerance for the convergence, by deafult set to NA.}

\item{verbose}{Logical, set to TRUE if you used tolerance criterion and you wish to see the messages about convergence, by deafult set to FALSE.}
}
\value{
A list with square root and inverse square root of \code{mat}.
}
\description{
A function for computing the square roots of a quadratic matrix using the Denman-Beavers algorithm. note that convergence is not guaranteed, even for matrices that do have square roots.
}
\examples{
# create a matrix
mat <- matrix(runif(9), 3, 3)

# taking a square root
sqrtmat(mat)
}

