% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{streval}
\alias{streval}
\title{A shortcut for evaluating an expression in a string format.}
\usage{
streval(x)
}
\arguments{
\item{x}{A character vector with a single element that contains an expression.}
}
\value{
Result of an evaluation of the expression contained in the string.
}
\description{
To evaluate an expression such as "1+1", we need to apply two function iteratively. streaval() function is a simple wrapper around eval() and parse() functions.
}
\examples{
# create some expression
x <- "1+2"

# when evaluated it should produce 3
streval(x)

# streval() function has the same output as:
eval(parse(text=x))
}
\seealso{
\code{\link{eval}}, \code{\link{parse}}
}

