% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.R
\name{tlag}
\alias{tlag}
\title{Creates a lagged variable of a vector.}
\usage{
tlag(x, d = 1, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric, character or logical vector that needs to be lagged.}

\item{d}{The number of lags (in units of observations). By default set to 1.}

\item{na.rm}{A logical that indicates whether NA elements should be removed from the output or not. By default set to FALSE.}
}
\value{
A numeric vector of the same length as \code{x} if \code{na.rm} is set to FALSE, otherwise it is of the same length as \code{x} without NA elements.
}
\description{
Creates a lagged variable of a numerical vector without requiring it to be of time series class.
}
\examples{
# create some numeric vector
x <- 1:10

# lag t-1, without removing the resulting NA elements and with removal
tslag(x, 1, FALSE)
tslag(x, 1, TRUE)
}

