% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{vecSum}
\alias{vecSum}
\title{Adding two vectors with special treatment of NA's.}
\usage{
vecSum(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A numeric vector of the same length as \code{x} and \code{y}.
}
\description{
Addition of two vectors that will ignore NA's if one of the two numbers is NA, but will return NA if both are NA.
}
\examples{
# create some vectors
x <- c(1,0,NA)
y <- c(1,NA,1)
z <- c(1,2,NA)

# first addition will ignore NA's while second will produce NA
vecSum(x, y)
vecSum(x, z)
}

