% Generated by roxygen2 (4.0.2): do not edit by hand
\name{changeClass}
\alias{changeClass}
\title{Change class of multiple variables in one go.}
\usage{
changeClass(df, newClass)
}
\arguments{
\item{df}{A data frame.}

\item{newClass}{A character vector which specifies desired class of each
variable in \code{df}.}
}
\value{
The data frame \code{df} where each column has a class as specified
in \code{newClass}.
}
\description{
Change class of multiple variables in one go.
}
\details{
The function also checks if one is converting a factor to numeric variable and it will make a correct
conversion without loss of digits.
}
\examples{
changeClass(mtcars, c("fac", rep("char",10)))
}
\seealso{
\code{\link{as.numeric}}, \code{\link{as.character}}, \code{\link{as.factor}}
 which this function wraps
}

