% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matlab.R
\name{repmat}
\alias{repmat}
\title{Replicate matrices.}
\usage{
repmat(mat, nrow = 1, ncol = 1)
}
\arguments{
\item{mat}{A matrix.}

\item{nrow}{A scalar that determines how many times \code{mat} will be replicated and added as new rows, each of the size of \code{mat}. Default value is 1.}

\item{ncol}{A scalar that determines how many times \code{mat} will be replicated and added as new columns, each of the size of \code{mat}. Default value is 1.}
}
\value{
A matrix where \code{mat} is copied \code{ncol} x \code{nrow} times and added as new columns as specified in \code{ncol} and new rows as specified in \code{nrow}.
}
\description{
A simple wrapper for \code{\link{matrix}} that emulates repmat from Matlab. The function behaves similar to \code{matrix} function when a single scalar is specified - it multiplies that scalar to fill out a matrix specified by nrow and ncol arguments. The repmat function takes the whole matrix and treats it as if it were a scalar, and replicates it as many times as needed to fill out a new matrix specified by \code{nrow} and \code{ncol} arguments.
}
\examples{
# create a simple matrix
mat <- matrix(1:4, 2, 2)

# now we take it as a block and replicate it to fill a 2x3 block matrix
repmat(mat, 2, 3)
}
\seealso{
\code{\link{matrix}}
}

